+++ src/urllib3/_async/connectionpool.py
@@ -6,6 +6,7 @@ import logging
 import queue
 import socket
 import sys
+import time
 import typing
 import warnings
 from asyncio import Task
@@ -15,6 +16,13 @@ from socket import timeout as SocketTimeout
 from types import TracebackType
 
 from .._collections import HTTPHeaderDict
+from .._constant import (
+    DEFAULT_BACKGROUND_WATCH_WINDOW,
+    DEFAULT_KEEPALIVE_DELAY,
+    DEFAULT_KEEPALIVE_IDLE_WINDOW,
+    MINIMAL_BACKGROUND_WATCH_WINDOW,
+    MINIMAL_KEEPALIVE_IDLE_WINDOW,
+)
 from .._request_methods import AsyncRequestMethods
 from .._typing import (
     _TYPE_ASYNC_BODY,
@@ -141,18 +149,53 @@ _blocking_errnos = {errno.EAGAIN, errno.EWOULDBLOCK}
 
 
 async def idle_conn_watch_task(
-    pool: AsyncHTTPConnectionPool, waiting_delay: float = 5.0
+    pool: AsyncHTTPConnectionPool,
+    waiting_delay: float | int = 5.0,
+    keepalive_delay: int | float | None = DEFAULT_KEEPALIVE_DELAY,
+    keepalive_idle_window: int | float | None = DEFAULT_KEEPALIVE_IDLE_WINDOW,
 ) -> None:
     """Discrete background task that monitor incoming data
     and dispatch message to registered callbacks."""
 
     while pool.pool is not None:
+        pool.num_background_watch_iter += 1
         await asyncio.sleep(waiting_delay)
         if pool.pool is None:
             return
         try:
             async for conn in pool.pool.iter_idle():
+                now = time.monotonic()
+                last_used = conn.last_used_at
+                idle_delay = now - last_used
+
+                # don't peek into conn that just became idle
+                # waste of resource.
+                if idle_delay < 1.0:
+                    continue
+
                 await conn.peek_and_react()
+
+                if keepalive_delay is not None and keepalive_idle_window is not None:
+                    connected_at = conn.connected_at
+
+                    if connected_at is not None:
+                        since_connection_delay = now - connected_at
+                        print(
+                            "Y",
+                            since_connection_delay <= keepalive_delay,
+                            since_connection_delay,
+                            keepalive_delay,
+                        )
+                        if since_connection_delay <= keepalive_delay:
+                            print(
+                                "X",
+                                idle_delay >= keepalive_idle_window,
+                                idle_delay,
+                                keepalive_idle_window,
+                            )
+                            if idle_delay >= keepalive_idle_window:
+                                await conn.ping()
+                                pool.num_pings += 1
         except AttributeError:
             return
 
@@ -204,6 +247,35 @@ class AsyncHTTPConnectionPool(AsyncConnectionPool, AsyncRequestMethods):
         A dictionary with proxy headers, should not be used directly,
         instead, see :class:`urllib3.AsyncProxyManager`
 
+    :param resolver:
+        A manual configuration to use for DNS resolution.
+        Can be a support DSN/str (e.g. "doh+cloudflare://") or a
+        :class:`urllib3.AsyncResolverDescription` or a list of DSN/str or
+        :class:`urllib3.AsyncResolverDescription`
+
+    :param happy_eyeballs:
+        Enable IETF Happy Eyeballs algorithm when trying to
+        connect by concurrently try multiple IPv4/IPv6 endpoints.
+        By default, tries at most 4 endpoints simultaneously.
+        You may specify an int that override this default.
+        Default, set to False.
+
+    :param background_watch_delay:
+        The window delay used by our discrete scheduler that run in
+        dedicated task to collect unsolicited incoming data and react
+        if necessary.
+
+    :param keepalive_delay:
+        The delay expressed in seconds on how long we should make sure
+        the connection is kept alive by sending pings to the remote peer.
+        Set it to None to void this feature.
+
+    :param keepalive_idle_window:
+        Immediately related to the parameter keepalive_delay.
+        This one, expressed in seconds, specify how long after
+        a connection is marked as idle we should send out a
+        ping to the remote peer.
+
     :param \\**conn_kw:
         Additional parameters are used to create fresh :class:`urllib3._async.connection.AsyncHTTPConnection`,
         :class:`urllib3._async.connection.AsyncHTTPSConnection` instances.
@@ -233,7 +305,9 @@ class AsyncHTTPConnectionPool(AsyncConnectionPool, AsyncRequestMethods):
         | AsyncBaseResolver
         | None = None,
         happy_eyeballs: bool | int = False,
-        background_watch_delay: int | float | None = 5.0,
+        background_watch_delay: int | float | None = DEFAULT_BACKGROUND_WATCH_WINDOW,
+        keepalive_delay: int | float | None = DEFAULT_KEEPALIVE_DELAY,
+        keepalive_idle_window: int | float | None = DEFAULT_KEEPALIVE_IDLE_WINDOW,
         **conn_kw: typing.Any,
     ):
         AsyncConnectionPool.__init__(self, host, port)
@@ -275,6 +349,9 @@ class AsyncHTTPConnectionPool(AsyncConnectionPool, AsyncRequestMethods):
         # These are mostly for testing and debugging purposes.
         self.num_connections = 0
         self.num_requests = 0
+        self.num_pings = 0
+        self.num_background_watch_iter = 0
+
         self.conn_kw = conn_kw
 
         if self.proxy:
@@ -341,8 +418,16 @@ class AsyncHTTPConnectionPool(AsyncConnectionPool, AsyncRequestMethods):
         )
 
         self.conn_kw["resolver"] = self._resolver
+        self.conn_kw["keepalive_delay"] = keepalive_delay
 
         self._background_watch_delay = background_watch_delay
+        self._keepalive_delay = keepalive_delay
+        self._keepalive_idle_window = keepalive_idle_window
+        if (
+            self._keepalive_idle_window is not None
+            and self._keepalive_idle_window < MINIMAL_KEEPALIVE_IDLE_WINDOW
+        ):
+            self._keepalive_idle_window = MINIMAL_KEEPALIVE_IDLE_WINDOW
         self._background_monitoring: asyncio.Task | None = None  # type: ignore[type-arg]
 
     @property
@@ -361,9 +446,15 @@ class AsyncHTTPConnectionPool(AsyncConnectionPool, AsyncRequestMethods):
         if (
             self._background_monitoring is None
             and self._background_watch_delay is not None
+            and self._background_watch_delay >= MINIMAL_BACKGROUND_WATCH_WINDOW
         ):
             self._background_monitoring = asyncio.create_task(
-                idle_conn_watch_task(self, self._background_watch_delay)
+                idle_conn_watch_task(
+                    self,
+                    self._background_watch_delay,
+                    self._keepalive_delay,
+                    self._keepalive_idle_window,
+                )
             )
 
         self.num_connections += 1
@@ -1879,9 +1970,18 @@ class AsyncHTTPSConnectionPool(AsyncHTTPConnectionPool):
         """
         if self.pool is None:
             raise ClosedPoolError(self, "Pool is closed")
-        if self._background_monitoring is None:
+        if (
+            self._background_monitoring is None
+            and self._background_watch_delay is not None
+            and self._background_watch_delay >= MINIMAL_BACKGROUND_WATCH_WINDOW
+        ):
             self._background_monitoring = asyncio.create_task(
-                idle_conn_watch_task(self)
+                idle_conn_watch_task(
+                    self,
+                    self._background_watch_delay,
+                    self._keepalive_delay,
+                    self._keepalive_idle_window,
+                )
             )
         self.num_connections += 1
         log.debug(