+++ src/urllib3/poolmanager.py
@@ -114,6 +114,8 @@ class PoolKey(typing.NamedTuple):
     key_key_data: str | bytes | None
     key_happy_eyeballs: bool | int | None
     key_background_watch_delay: float | int | None
+    key_keepalive_delay: float | int | None
+    key_keepalive_idle_window: float | int | None
 
 
 @functools.lru_cache(maxsize=8)